refactor: 클래스 패키지 이동
- CustomMoveStrategy는 테스트에서만 사용하므로 test 패키지로 이동
- TryCount는 Controller에서 사용하므로 controller로 이동
