+++ src/test/java/racingcar/controller/TryCountTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.controller;
 
 import static org.assertj.core.api.Assertions.*;
